requirements.txt
absl-py==2.1.0
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
archspec @ file:///home/conda/feedstock_root/build_artifacts/archspec_1699370045702/work
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
black==24.2.0
boltons @ file:///home/conda/feedstock_root/build_artifacts/boltons_1677499911949/work
boto3==1.35.65
botocore==1.35.65
Brotli @ file:///Users/runner/miniforge3/conda-bld/brotli-split_1695989934239/work
bs4==0.0.1
cachetools==5.3.2
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1700303426725/work/certifi
cffi @ file:///Users/runner/miniforge3/conda-bld/cffi_1696001737800/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1698833585322/work
chroma-hnswlib==0.7.3
chromadb==0.4.17
click==8.1.7
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
coloredlogs==15.0.1
conda @ file:///Users/runner/miniforge3/conda-bld/conda_1701731166956/work
conda-libmamba-solver @ file:///home/conda/feedstock_root/build_artifacts/conda-libmamba-solver_1698967720775/work/src
conda-package-handling @ file:///home/conda/feedstock_root/build_artifacts/conda-package-handling_1691048088238/work
conda_package_streaming @ file:///home/conda/feedstock_root/build_artifacts/conda-package-streaming_1691009212940/work
contourpy==1.2.0
cryptography==41.0.5
cycler==0.12.1
dataclasses-json==0.6.2
dateparser==1.1.8
defusedxml==0.7.1
Deprecated==1.2.14
distro @ file:///home/conda/feedstock_root/build_artifacts/distro_1704321475663/work
docstring-to-markdown==0.15
exceptiongroup==1.1.3
fastapi==0.104.1
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.47.2
frozenlist==1.4.0
fsspec==2023.10.0
gensim==4.3.2
google-auth==2.23.4
googleapis-common-protos==1.61.0
gpt4all==2.0.2
grpcio==1.59.2
h11==0.14.0
httpcore==1.0.7
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.17.3
humanfriendly==10.0
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1663625384323/work
importlib-metadata==6.8.0
importlib-resources==6.1.1
insta-scrape==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonpatch @ file:///home/conda/feedstock_root/build_artifacts/jsonpatch_1695536281965/work
jsonpointer @ file:///Users/runner/miniforge3/conda-bld/jsonpointer_1695397382064/work
kiwisolver==1.4.5
kubernetes==28.1.0
langchain==0.0.335
langchain-core==0.2.43
langchain-pinecone==0.1.2
langsmith==0.1.144
libmambapy @ file:///Users/runner/miniforge3/conda-bld/mamba-split_1698738490091/work/libmambapy
lxml==5.2.2
lz4==4.3.3
mamba @ file:///Users/runner/miniforge3/conda-bld/mamba-split_1698738490091/work/mamba
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.2
mdurl==0.1.2
menuinst @ file:///Users/runner/miniforge3/conda-bld/menuinst_1705068367719/work
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.1
oauthlib==3.2.2
onnxruntime==1.16.2
openai==0.28.1
opentelemetry-api==1.21.0
opentelemetry-exporter-otlp-proto-common==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-proto==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-semantic-conventions==0.42b0
orjson==3.10.11
osmium==3.6.0
overrides==7.4.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1696202382185/work
pandas==2.1.1
parso==0.8.3
pathspec==0.12.1
pdfminer.six==20221105
Pillow==9.3.0
pinecone==5.4.0
pinecone-client==4.1.2
pinecone-plugin-inference==2.0.1
pinecone-plugin-interface==0.0.7
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1701708255999/work
plotly==5.17.0
pluggy @ file:///home/conda/feedstock_root/build_artifacts/pluggy_1693086607691/work
posthog==3.0.2
protobuf==4.25.0
protoc-gen-openapiv2==0.0.1
pulsar-client==3.3.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycosat @ file:///Users/runner/miniforge3/conda-bld/pycosat_1696355969704/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
pydantic==1.10.13
Pygments==2.16.1
PyMuPDF==1.23.6
PyMuPDFb==1.23.6
pyparsing==3.1.1
PyPDF2==1.26.0
PyPika==0.48.9
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1661604839144/work
python-dateutil==2.8.2
python-dotenv==1.0.0
python-lsp-black==2.0.0
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.10.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0
redis==5.0.8
regex==2023.10.3
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1684774241324/work
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.6.0
rsa==4.9
ruamel.yaml @ file:///Users/runner/miniforge3/conda-bld/ruamel.yaml_1699007390921/work
ruamel.yaml.clib @ file:///Users/runner/miniforge3/conda-bld/ruamel.yaml.clib_1695996997220/work
s3transfer==0.10.3
safetensors==0.4.0
schedule==1.1.0
scikit-learn==1.3.2
scipy==1.11.3
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
soupsieve==2.5
spotipy==2.24.0
SQLAlchemy==2.0.23
starlette==0.27.0
sympy==1.12
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
threadpoolctl==3.2.0
tiktoken==0.5.1
tokenizers==0.14.1
tomli==2.0.1
torch==2.0.1
torchvision==0.15.2
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1691671248568/work
transformers==4.35.0
truststore @ file:///home/conda/feedstock_root/build_artifacts/truststore_1694154605758/work
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
ujson==5.9.0
urllib3==1.26.18
uvicorn==0.24.0.post1
uvloop==0.19.0
validate-email==1.3
watchfiles==0.21.0
websocket-client==1.6.4
websockets==12.0
Werkzeug==3.0.1
wrapt==1.16.0
yarl==1.9.2
zipp==3.17.0
zstandard==0.22.0
